bench: rename benchmark project to MicroBench2
Make it a general lean BenchmarkDotNet host (only the project under test
plus BenchmarkDotNet, no heavier third-party deps like MicroBenchmark)
so future minimal-dependency micro-benchmarks can be added alongside the
base64url encoder one.
